Remove a huge number of kernel modules
Many of these were deleted as part of building the root filesystem
anyway, so removing them speeds up build time. Others that I judged not
to be necessary were also removed.

This was a relatively conservative cull, if building a kernel image only
to run on top of virtualbox a much more aggressive cull could be done.
It is however possible that some drivers used by boot2docker
configurations in the wild have been removed.